Scatter: move gradient logic inside Gradient class
