+++ src/ScottPlot5/ScottPlot5/Plottables/Scatter.cs
@@ -126,65 +126,6 @@ public class Scatter(IScatterSource data) : IPlottable, IHasLine, IHasMarker, IH
 
     public IEnumerable<LegendItem> LegendItems => LegendItem.Single(this, LegendText, MarkerStyle, LineStyle);
 
-    private Gradient CreateAxisGradient(RenderPack rp)
-    {
-        AxisLimits dataLimits = Data.GetLimits();
-        double min = FillGradientDirection == GradientDirection.Horizontal ? (double)Axes.XAxis.GetCoordinate(rp.DataRect.Left, rp.DataRect) : dataLimits.Bottom;
-        double max = FillGradientDirection == GradientDirection.Horizontal ? (double)Axes.XAxis.GetCoordinate(rp.DataRect.Right, rp.DataRect) : dataLimits.Top;
-
-        var sortedColorPositions = ColorPositions.OrderBy(cp => cp.Position).ToList();
-
-        Color interpolate(double val)
-        {
-            if (sortedColorPositions.Count == 0)
-                return Colors.Black;
-
-            if (val <= sortedColorPositions.First().Position)
-                return sortedColorPositions.First().Color;
-
-            if (val >= sortedColorPositions.Last().Position)
-                return sortedColorPositions.Last().Color;
-
-            int upperIndex = sortedColorPositions.FindIndex(cp => cp.Position >= val);
-            int lowerIndex = upperIndex - 1;
-
-            var lower = sortedColorPositions[lowerIndex];
-            var upper = sortedColorPositions[upperIndex];
-
-            double fraction = (val - lower.Position) / (upper.Position - lower.Position);
-            return lower.Color.InterpolateRgb(upper.Color, fraction);
-        }
-
-        var stops = sortedColorPositions
-            .Select(cp => cp.Position)
-            .Where(p => p >= min && p <= max)
-            .Concat(new[] { min, max })
-            .Distinct()
-            .OrderBy(p => p)
-            .ToList();
-
-        float GetFraction(double position)
-        {
-            if (max == min) return 0;
-            return (float)((position - min) / (max - min));
-        }
-
-        var (alignmentStart, alignmentEnd) = FillGradientDirection switch
-        {
-            GradientDirection.Horizontal => (Alignment.MiddleLeft, Alignment.MiddleRight),
-            GradientDirection.Vertical => (Alignment.LowerCenter, Alignment.UpperCenter),
-            _ => throw new NotImplementedException()
-        };
-
-        return new Gradient(GradientType.Linear)
-        {
-            AlignmentStart = alignmentStart,
-            AlignmentEnd = alignmentEnd,
-            ColorPositions = stops.Select(GetFraction).ToArray(),
-            Colors = stops.Select(interpolate).ToArray()
-        };
-    }
-
     public virtual void Render(RenderPack rp)
     {
         var coordinates = Data.GetScatterPoints();
@@ -217,10 +158,10 @@ public class Scatter(IScatterSource data) : IPlottable, IHasLine, IHasMarker, IH
 
             if (ColorPositions.Count > 0)
             {
-                fs.Hatch = CreateAxisGradient(rp);
+                fs.Hatch = Gradient.CreateAxisGradient(rp, Data.GetLimits(), FillGradientDirection, Axes, ColorPositions);
             }
-            PixelRect dataPxRect = new(markerPixels);
 
+            PixelRect dataPxRect = new(markerPixels);
             PixelRect rect = new(linePixels);
             float yValuePixel = Axes.YAxis.GetPixel(FillYValue + OffsetY, rp.DataRect);
