+++ src/ScottPlot5/ScottPlot5/Primitives/Gradient.cs
@@ -1,3 +1,5 @@
+using ScottPlot.Primitives;
+
 namespace ScottPlot;
 
 public class Gradient(GradientType gradientType = GradientType.Linear) : IHatch
@@ -85,6 +87,64 @@ public class Gradient(GradientType gradientType = GradientType.Linear) : IHatch
                 colorPos: ColorPositions,
                 mode: TileMode
                 ),
+        };
+    }
+
+    public static Gradient CreateAxisGradient(RenderPack rp, AxisLimits limits, GradientDirection direction, IAxes axes, List<GradientColorPosition> colorPositions)
+    {
+        double min = direction == GradientDirection.Horizontal ? (double)axes.XAxis.GetCoordinate(rp.DataRect.Left, rp.DataRect) : limits.Bottom;
+        double max = direction == GradientDirection.Horizontal ? (double)axes.XAxis.GetCoordinate(rp.DataRect.Right, rp.DataRect) : limits.Top;
+
+        var sortedColorPositions = colorPositions.OrderBy(cp => cp.Position).ToList();
+
+        Color interpolate(double val)
+        {
+            if (sortedColorPositions.Count == 0)
+                return ScottPlot.Colors.Black;
+
+            if (val <= sortedColorPositions.First().Position)
+                return sortedColorPositions.First().Color;
+
+            if (val >= sortedColorPositions.Last().Position)
+                return sortedColorPositions.Last().Color;
+
+            int upperIndex = sortedColorPositions.FindIndex(cp => cp.Position >= val);
+            int lowerIndex = upperIndex - 1;
+
+            var lower = sortedColorPositions[lowerIndex];
+            var upper = sortedColorPositions[upperIndex];
+
+            double fraction = (val - lower.Position) / (upper.Position - lower.Position);
+            return lower.Color.InterpolateRgb(upper.Color, fraction);
+        }
+
+        var stops = sortedColorPositions
+            .Select(cp => cp.Position)
+            .Where(p => p >= min && p <= max)
+            .Concat(new[] { min, max })
+            .Distinct()
+            .OrderBy(p => p)
+            .ToList();
+
+        float GetFraction(double position)
+        {
+            if (max == min) return 0;
+            return (float)((position - min) / (max - min));
+        }
+
+        var (alignmentStart, alignmentEnd) = direction switch
+        {
+            GradientDirection.Horizontal => (Alignment.MiddleLeft, Alignment.MiddleRight),
+            GradientDirection.Vertical => (Alignment.LowerCenter, Alignment.UpperCenter),
+            _ => throw new NotImplementedException()
+        };
+
+        return new Gradient(GradientType.Linear)
+        {
+            AlignmentStart = alignmentStart,
+            AlignmentEnd = alignmentEnd,
+            ColorPositions = stops.Select(GetFraction).ToArray(),
+            Colors = stops.Select(interpolate).ToArray()
         };
     }
 }
